Add an `avoid` dictionary, plus improve spellcheck output.
The `avoid` dictionary contains words that are correctly spelled
in the content-style-guide, but that are uncommon words that should be
flagged as spelling errors everywhere else.

The spellcheck script's output now begins with misspellings and where
they occur, and a new section with the files and their misspellings
(helpful if you have many misspellings, and are fixing file-by-file).